Explicitly skipping python 2 build for prismatic sub-package.
Also added python dependency to host section for the
prismatic sub-package (for linux only).

The skip change ensures that we don't build a spurious python 2
package, when only python 3 is supported by pyprismatic. And the
addition to host ensures that the prismatic subpackage is built
for all versions of python 3 (the dependence on the pyprismatic
subpackage is not enough here).